restructured TimePeriod folders
